Extract `replica_set/update.rs`
